+++ src/main/java/com/eghm/websocket/mapper/UserMapper.java
@@ -23,7 +23,7 @@ public interface UserMapper {
      *
      * @param id 条件 
      */
-    User selectByPrimaryKey(Integer id);
+    User selectByPrimaryKey(Long id);
 
     /**
      * 根据主键来更新部分数据库记录
